# Add compression and decompression benchmarks