Refs https://github.com/clue/reactphp-buzz/pull/149, https://github.com/clue/reactphp-ndjson/pull/16 and others